Separate the three main formats into individual files
